Improved backup generation for init.sh

+++ init.sh
@@ -1,13 +1,12 @@
 #! /bin/bash
 
-FILES=( 'zshrc' 'zsh_aliases' 'zsh_functions' 'vim' 'vimrc' 'xprofile' 'Xmodmap' 'SciTEUser.properties' 'gitconfig' 'templates' 'tmux.conf' 'tmux.zsh' )
+FILES=( 'zshrc' 'zsh_aliases' 'zsh_functions' 'vim' 'vimrc' 'xprofile' 'Xmodmap' 'SciTEUser.properties' 'gitconfig' 'templates' 'tmux.conf' 'tmux.zsh' 'i3' )
 
 cd $HOME
 
 for file in ${FILES[@]}; do
-    if [ -e ".${file}" ]; then
-        mv ".${file}" ".${file}.bak"
-    fi
+    [[ -L ".${file}" ]] && rm ".${file}"
+    [[ -f ".${file}" ]] && mv ".${file}" ".${file}.bak"
 
     ln -s ".dotfiles/${file}" ".${file}"
 done
